backend/requirements.txt
fastapi
pydantic
openai>=1.0.0
anthropic>=0.8.0
python-dotenv
pytest
pytest-asyncio
uvicorn
python-multipart
